third_party/WebKit/LayoutTests/fast/canvas/canvas-createImageBitmap-invalid-args.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Ensure correct behavior of createImageBitmap for invalid inputs.");
window.jsTestIsAsync = true;

var InvalidStateError = "InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable.";
var TypeError = "TypeError: Type error";
var IndexSizeError = "IndexSizeError: Index or size was negative, or greater than the allowed value.";

var image;
var testBitmap; // an ImageBitmap that is uncropped. We use this to test createImageBitmap(testBitmap)
var d;          // image.imageData
var blob;
var invalidBlob;

// Draw to an auxillary canvas.
var aCanvas = document.createElement("canvas");
aCanvas.setAttribute("width", "200");
aCanvas.setAttribute("height", "200");
var aCtx = aCanvas.getContext("2d");

image = new Image();
image.onload = imageLoaded;

// Before image loads
shouldThrow("createImageBitmap(image)", "InvalidStateError");
image.src = aCanvas.toDataURL();

video = document.createElement("video");
video.addEventListener("canplaythrough", videoLoaded, false);

// Before video loads
shouldThrow("createImageBitmap(video)", "InvalidStateError");
video.src = "../../compositing/resources/video.ogv";

var imageLoaded = false;
var videoLoaded = false;
var imageBitmapLoaded = false;
var blobLoaded = false;
var invalidBlobLoaded = false;

function imageLoaded() {
    createImageBitmap(image).then(imageBitmapLoadedCallback, function() {
        testFailed("Promise was rejected.");
        finishJSTest();
    });
    d = aCtx.getImageData(0, 0, 200, 200);
    imageLoaded = true;
    loaded();
}

function videoLoaded() {
    videoLoaded = true;
    loaded();
}

function imageBitmapLoadedCallback(imageBitmap) {
    testBitmap = imageBitmap;
    imageBitmapLoaded = true;
    loaded();
}

var xhr = new XMLHttpRequest();
xhr.open("GET", 'resources/pattern.png');
xhr.responseType = 'blob';
xhr.send();
xhr.onload = function() {
    blob = xhr.response;
    blobLoaded = true;
    loaded();
}

var xhr2 = new XMLHttpRequest();
xhr2.open("GET", 'resources/repaint.js');
xhr2.responseType = 'blob';
xhr2.send();
xhr2.onload = function() {
    invalidBlob = xhr2.response;
    invalidBlobLoaded = true;
    loaded();
}

function loaded() {
    if (imageLoaded && videoLoaded && imageBitmapLoaded && blobLoaded && invalidBlobLoaded) {
        shouldThrow("createImageBitmap(undefined)", "TypeError");

        shouldThrow("createImageBitmap(image, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(image, 0, 0, 0, 10)", "IndexSizeError");

        shouldThrow("createImageBitmap(video, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(video, 0, 0, 0, 10)", "IndexSizeError");

        shouldThrow("createImageBitmap(aCanvas, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(aCanvas, 0, 0, 0, 10)", "IndexSizeError");

        shouldThrow("createImageBitmap(d, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(d, 0, 0, 0, 10)", "IndexSizeError");

        shouldThrow("createImageBitmap(aCtx, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(aCtx, 0, 0, 0, 10)", "IndexSizeError");

        shouldThrow("createImageBitmap(testBitmap, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(testBitmap, 0, 0, 0, 10)", "IndexSizeError");

        shouldThrow("createImageBitmap(blob, 0, 0, 10, 0)", "IndexSizeError");
        shouldThrow("createImageBitmap(blob, 0, 0, 0, 10)", "IndexSizeError");

        createImageBitmap(invalidBlob).then(function() {
            testFailed("Invalid blob fulfilled.");
            finishJSTest();
        }, function() {
            testPassed("createImageBitmap(invalidBlob) was rejected.");
            finishJSTest();
        });
    }
}

</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
